switch to ssize_t

+++ get_next_line.c
@@ -6,7 +6,7 @@
 /*   By: vvagapov <vvagapov@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/26 15:53:55 by vvagapov          #+#    #+#             */
-/*   Updated: 2022/12/27 15:24:19 by vvagapov         ###   ########.fr       */
+/*   Updated: 2022/12/27 15:35:05 by vvagapov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,9 @@
 
 // Allocates enough memory for new result, copies old result and additional
 // chars from buffer to new memory, frees memory from old result
-static char	*append_line(char *old_res, char *buf, int start, int len)
+static char	*append_line(char *old_res, char *buf, ssize_t start, ssize_t len)
 {
-	int		i;
+	ssize_t		i;
 	char	*res;
 
 	res = (char *)malloc(sizeof(char)
@@ -51,7 +51,7 @@ static char	*append_line(char *old_res, char *buf, int start, int len)
 // If 1 is returned, data was read to buffer and the process can continue
 static int	read_to_buf(int fd, char **res, char *buf)
 {
-	int	ret;
+	ssize_t	ret;
 
 	ret = read(fd, buf, BUFFER_SIZE);
 	if (ret == 0 && (**res))
@@ -70,7 +70,7 @@ static int	read_to_buf(int fd, char **res, char *buf)
 // 		and therefore it's time to add the last chunk to result and return it.
 // line_start is set to the next character after newline,
 // line_len is set to 0 in preparation for next get_next_line() call.
-static char	*finish_line(char *res, char *buf, int *line_start, int *line_len)
+static char	*finish_line(char *res, char *buf, ssize_t *line_start, ssize_t *line_len)
 {
 	res = append_line(res, buf, *line_start, *line_len);
 	*line_start += *line_len;
@@ -83,8 +83,8 @@ static char	*finish_line(char *res, char *buf, int *line_start, int *line_len)
 // It returns 0 in case or error (malloc fail), returns 1 if success.
 // line_start and line_len values are also reset
 //		in preparation for next buffer read.
-static int	add_buf_to_res(char **res, char *buf, int *line_start,
-		int *line_len)
+static int	add_buf_to_res(char **res, char *buf, ssize_t *line_start,
+		ssize_t *line_len)
 {
 	*res = append_line(*res, buf, *line_start, BUFFER_SIZE - *line_start);
 	if (!*res)
@@ -109,10 +109,10 @@ static int	add_buf_to_res(char **res, char *buf, int *line_start,
 //		otherwise NULL is returned.
 char	*get_next_line(int fd)
 {
-	char		*res;
-	static char	buf[BUFFER_SIZE + 1];
-	static int	line_start = 0;
-	static int	line_len = 0;
+	char			*res;
+	static char		buf[BUFFER_SIZE + 1];
+	static ssize_t	line_start = 0;
+	static ssize_t	line_len = 0;
 
 	if (fd < 0 || BUFFER_SIZE < 1 || !init_res(&res))
 		return (NULL);
