+++ get_next_line_utils.c
@@ -6,7 +6,7 @@
 /*   By: vvagapov <vvagapov@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/26 15:54:32 by vvagapov          #+#    #+#             */
-/*   Updated: 2022/12/27 15:00:54 by vvagapov         ###   ########.fr       */
+/*   Updated: 2022/12/27 15:32:53 by vvagapov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 // Looks for a newline character in buffer starting from start_index.
 // Returns the index of the character folowing the newline.
 // Returns 0 is no newline was found.
-int	find_newline(char *buf, int start_index)
+ssize_t	find_newline(char *buf, ssize_t start_index)
 {
 	while (buf[start_index])
 	{
@@ -27,9 +27,9 @@ int	find_newline(char *buf, int start_index)
 }
 
 // Returns the length of the string s
-size_t	ft_strlen(const char *s)
+ssize_t	ft_strlen(const char *s)
 {
-	size_t	res;
+	ssize_t	res;
 
 	res = 0;
 	while (s[res])
